ai-meeting-summary/requirements.txt
openai-whisper
transformers
torch
flask
